pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-fitrl"
version = "0.1.0"
description = "AI Life Optimization environment for OpenEnv — 7-day episode with fitness, work, and recovery phases"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "pydantic>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m fitrl.server.app
server = "fitrl.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["fitrl", "fitrl.server"]
package-dir = { "fitrl" = ".", "fitrl.server" = "server" }